automotive-backend/requirements.txt
Django==5.1
django-rest-framework
django-cors-headers
